refactor: 영화 목록 조회 기능 1차 리팩토링
