+++ src/constans.ts
@@ -1,0 +1,6 @@
+export const FETCH_OPTION = {
+  headers: {
+    accept: "application/json",
+    Authorization: `Bearer ${import.meta.env.VITE_ACCESS_TOKEN}`,
+  }
+}
